Jobs sandbox and readme
