+++ samples/deprecated/fleetprovisioning_mqtt5.py
