+++ samples/shadow.md
@@ -90,7 +90,7 @@ If the venv does not yet have the device SDK installed, install it:
 python3 -m pip install awsiotsdk
 ```
 
-Assuming you are in the SDK root directory, you can now run the shadow sanbox sample:
+Assuming you are in the SDK root directory, you can now run the shadow sandbox sample:
 
 ``` sh
 python3 samples/shadow.py --cert <path to certificate> --key <path to private key> --endpoint <account-specific broker endpoint> --thing <thing name>
